+++ include/mlir-ext/Dialect/CommonIR/IR/CommonIRDialect.h
@@ -1,4 +1,4 @@
-//===- TileIRDialect.h - MLIR TileIR dialect --------------------*- C++ -*-===//
+//===- CommonIRDialect.h - MLIR CommonIR dialect --------------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,7 +6,7 @@
 //
 //===----------------------------------------------------------------------===//
 //
-// This file defines the TileIR dialect in MLIR, containing operations that
+// This file defines the CommonIR dialect in MLIR, containing operations that
 // explicitly encode memory hierarchy, engine binding, data layout, and
 // synchronization for deterministic 1:1 lowering to HIVM.
 //
@@ -21,18 +21,18 @@
 
 #include "triton/Dialect/Triton/IR/Dialect.h"
 
-#include "mlir-ext/Dialect/TileIR/IR/TileIRDialect.h.inc"
+#include "mlir-ext/Dialect/CommonIR/IR/CommonIRDialect.h.inc"
 
-#include "mlir-ext/Dialect/TileIR/IR/TileIREnums.h.inc"
+#include "mlir-ext/Dialect/CommonIR/IR/CommonIREnums.h.inc"
 
 #define GET_ATTRDEF_CLASSES
-#include "mlir-ext/Dialect/TileIR/IR/TileIROpsAttrDefs.h.inc"
+#include "mlir-ext/Dialect/CommonIR/IR/CommonIROpsAttrDefs.h.inc"
 
 #define GET_TYPEDEF_CLASSES
-#include "mlir-ext/Dialect/TileIR/IR/TileIRTypes.h.inc"
+#include "mlir-ext/Dialect/CommonIR/IR/CommonIRTypes.h.inc"
 
 #define GET_OP_CLASSES
-#include "mlir-ext/Dialect/TileIR/IR/TileIROps.h.inc"
+#include "mlir-ext/Dialect/CommonIR/IR/CommonIROps.h.inc"
 
 namespace mlir::triton::tile {} // namespace mlir::triton::tile
