+++ include/mlir-ext/Dialect/CommonIR/Transforms/Passes.td
@@ -3,14 +3,14 @@
 
 include "mlir/Pass/PassBase.td"
 
-def TileIRInferLayout : Pass<"tile-infer-layout", "mlir::ModuleOp"> {
-  let summary = "Infer default TileIR layout attributes";
+def CommonIRInferLayout : Pass<"tile-infer-layout", "mlir::ModuleOp"> {
+  let summary = "Infer default CommonIR layout attributes";
   let description = [{
-    Fills missing TileIR layout attributes with a conservative ND layout.
+    Fills missing CommonIR layout attributes with a conservative ND layout.
     More target-specific layouts, such as GPU shared-memory swizzles or dot
     operand layouts, can be refined by later target-aware passes.
   }];
-  let constructor = "mlir::triton::tile::createTileIRInferLayoutPass()";
+  let constructor = "mlir::triton::tile::createCommonIRInferLayoutPass()";
 }
 
 #endif // TILE_IR_TRANSFORMS_PASSES
