+++ lib/Dialect/CMakeLists.txt
@@ -2,7 +2,7 @@ add_subdirectory(TritonTilingExt)
 add_subdirectory(TPtr)
 add_subdirectory(MathExt)
 add_subdirectory(TritonStructured)
-add_subdirectory(TileIR)
+add_subdirectory(CommonIR)
 if (FLIR_BUILD_INCUBATED)
   add_subdirectory(TritonStructuredIncubated)
 endif()
