+++ lib/Dialect/CommonIR/IR/CommonIRAttrs.cpp
@@ -1,4 +1,4 @@
-//===- TileIRAttrs.cpp - TileIR attribute implementations -----------------===//
+//===- CommonIRAttrs.cpp - CommonIR attribute implementations -----------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,11 +6,11 @@
 //
 //===----------------------------------------------------------------------===//
 //
-// This file implements TileIR attribute custom methods.
+// This file implements CommonIR attribute custom methods.
 //
 //===----------------------------------------------------------------------===//
 
-#include "mlir-ext/Dialect/TileIR/IR/TileIRDialect.h"
+#include "mlir-ext/Dialect/CommonIR/IR/CommonIRDialect.h"
 
 using namespace mlir;
 using namespace mlir::triton::tile;
