+++ lib/Dialect/CommonIR/Transforms/CMakeLists.txt
@@ -1,0 +1,11 @@
+add_triton_library(CommonIRTransforms
+  InferLayout.cpp
+
+  DEPENDS
+  CommonIRTransformsPassIncGen
+
+  LINK_LIBS PUBLIC
+  MLIRIR
+  MLIRPass
+  CommonIRIR
+)
