--- lib/Dialect/TileIR/IR/CMakeLists.txt
@@ -1,15 +1,0 @@
-add_triton_library(TileIRIR
-  TileIRAttrs.cpp
-  TileIRDialect.cpp
-  TileIROps.cpp
-
-  DEPENDS
-  TileIRTableGen
-  TileIRAttrDefsIncGen
-  TileIRTypesIncGen
-
-  LINK_LIBS PUBLIC
-  MLIRIR
-  MLIRLLVMDialect
-  TritonIR
-)
