--- lib/Dialect/TileIR/Transforms/CMakeLists.txt
@@ -1,11 +1,0 @@
-add_triton_library(TileIRTransforms
-  InferLayout.cpp
-
-  DEPENDS
-  TileIRTransformsPassIncGen
-
-  LINK_LIBS PUBLIC
-  MLIRIR
-  MLIRPass
-  TileIRIR
-)
